refactor: centralize prompt configuration
